chore(Projects): Rework the project rows hook
