+++ src/pages/ProjectsPage/hooks/index.ts
@@ -7,4 +7,4 @@ export * from './useProjectFilters'
 export * from './useProjectGrouping'
 export * from './useBuildProjectFilterOptions'
 export * from './utils/filterProjects'
-export * from './useProjectGroupedRows'
+export * from './useProjectTableRows'
